Refactor login view out of AccountView
